# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ReMynd Assessment Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cases/{caseId}/scores:
    get:
      operationId: getCaseScores
      tags: [scores]
      summary: Get all scores for a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Score"
  /cases/{caseId}/scores/calculate:
    post:
      operationId: calculateScores
      tags: [scores]
      summary: Calculate/recalculate scores for a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Calculated scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Score"
  /cases/{caseId}/scores/manual:
    post:
      operationId: submitManualScore
      tags: [scores]
      summary: Submit manual score (external tools)
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualScoreRequest"
      responses:
        "201":
          description: Score submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Score"

  # ─── Reports ──────────────────────────────────────────────────────────────
...
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/login:
    post:
      operationId: login
      tags: [users]
      summary: Login with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, assessment_lead, psychometrician]
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, createdAt]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]

    # ─── Case ────────────────────────────────────────────────────────────────
...
    Score:
      type: object
      properties:
        id:
          type: string
        caseId:
          type: string
        toolId:
          type: string
        toolName:
          type: string
        respondentType:
          type: string
        rawScore:
          type: number
          nullable: true
        domainScores:
          type: object
          additionalProperties:
            type: number
        normalizedScores:
          type: object
          additionalProperties:
            type: number
        agreementIndex:
          type: number
          nullable: true
        hasHighDiscrepancy:
          type: boolean
        isManual:
          type: boolean
        notes:
          type: string
          nullable: true
        generatedAt:
          type: string
          format: date-time
      required:
        - id
        - caseId
        - toolId
        - toolName
        - respondentType
        - domainScores
        - normalizedScores
        - hasHighDiscrepancy
        - isManual
        - generatedAt
    ManualScoreRequest:
      type: object
      properties:
        toolId:
          type: string
        toolName:
          type: string
        respondentType:
          type: string
        rawScore:
          type: number
        domainScores:
          type: object
          additionalProperties:
            type: number
        notes:
          type: string
      required: [toolId, toolName, respondentType, rawScore, domainScores]

    # ─── Report ───────────────────────────────────────────────────────────────
...